+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Add `display_name`, `ticker`, `coin_hours_display_name`, `coin_hours_ticker`, `explorer_url` to the `/health` endpoint response
+- Add new wallet type `collection` for wallets that are an arbitrary collection of private keys, rather than generated from a seed
+- `cli addPrivateKey` will only work on a `collection` type wallet. Create one with `cli walletCreate -t collection`
+- Don't print the wallet in the terminal after `cli encryptWallet` or `cli decryptWallet`
 
 ### Removed